requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-dotenv
ultralytics
opencv-python-headless
psycopg2-binary
python-dotenv